+++ templates/wp-home.php
@@ -3,7 +3,7 @@
  * The template displaying the posts-overview
  *
  * @author      Flurin Dürst
- * @version     1.4
+ * @version     1.4.1
  * @since       WPSeed 0.18
  */
 
@@ -11,7 +11,7 @@
 
 <? get_header(); ?>
 
-<!--- content » blog home --->
+<!-- content » blog home -->
 
   <div class="content blog">