+++ templates/wp-page.php
@@ -1,14 +1,14 @@
 <?
 /**
  * @author      Flurin Dürst
- * @version     1.2
+ * @version     1.2.1
  * @since       WPSeed 0.1
  */
 ?>
 
 <? get_header(); ?>
 
-<!--- content » page --->
+<!-- content » page -->
 
   <div class="content page">