+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Head
+* adjustments for w3c compability
+* use page as index
+
 ## 0.11.0 | 05.07.2016 | «CleanFlex»
 * huge cleanup/simplification of css and templates
 * use flexbox as template default